add ModelParallel class
