+++ src/models.py
@@ -1,16 +1,90 @@
 import torch
 from torch.autograd import Variable
 from torch import optim
-from .steppy.pytorch.models import Model
+from toolkit.pytorch_transformers.models import Model
 from toolkit.pytorch_transformers.callbacks import CallbackList, TrainingMonitor, ModelCheckpoint, \
     ExperimentTiming, ExponentialLRScheduler, EarlyStopping, NeptuneMonitor, ValidationMonitor
-from .steppy.pytorch.parallel import DataParallelCriterion
 
+from .parallel import DataParallelCriterion, DataParallelModel as DataParallel
 from .callbacks import NeptuneMonitorDetection, ValidationMonitorDetection
 from .retinanet import RetinaNet, RetinaLoss
 
 
-class Retina(Model):
+class ModelParallel(Model):
+    def fit(self, datagen, validation_datagen=None):
+        self._initialize_model_weights()
+
+        self.model = DataParallel(self.model)
+
+        if torch.cuda.is_available():
+            self.model = self.model.cuda()
+
+        self.callbacks.set_params(self, validation_datagen=validation_datagen)
+        self.callbacks.on_train_begin()
+
+        batch_gen, steps = datagen
+        for epoch_id in range(self.training_config['epochs']):
+            self.callbacks.on_epoch_begin()
+            for batch_id, data in enumerate(batch_gen):
+                self.callbacks.on_batch_begin()
+                metrics = self._fit_loop(data)
+                self.callbacks.on_batch_end(metrics=metrics)
+                if batch_id == steps:
+                    break
+            self.callbacks.on_epoch_end()
+            if self.callbacks.training_break():
+                break
+        self.callbacks.on_train_end()
+        return self
+
+    def _fit_loop(self, data):
+        X = data[0]
+        targets_tensors = data[1:]
+
+        if torch.cuda.is_available():
+            X = Variable(X).cuda()
+            targets_var = []
+            for target_tensor in targets_tensors:
+                targets_var.append(Variable(target_tensor).cuda())
+        else:
+            X = Variable(X)
+            targets_var = []
+            for target_tensor in targets_tensors:
+                targets_var.append(Variable(target_tensor))
+
+        self.optimizer.zero_grad()
+        outputs_batch = self.model(X)
+        partial_batch_losses = {}
+
+        if len(self.output_names) == 1:
+            for (name, loss_function, weight), target in zip(self.loss_function, targets_var):
+                batch_loss = loss_function(outputs_batch, target) * weight
+        else:
+            for (name, loss_function, weight), output, target in zip(self.loss_function, outputs_batch, targets_var):
+                partial_batch_losses[name] = loss_function(output, target) * weight
+            batch_loss = sum(partial_batch_losses.values())
+        partial_batch_losses['sum'] = batch_loss
+        batch_loss.backward()
+        self.optimizer.step()
+
+        return partial_batch_losses
+
+    def load(self, filepath):
+        self.model.eval()
+
+        if not isinstance(self.model, DataParallel):
+            self.model = DataParallel(self.model)
+
+        if torch.cuda.is_available():
+            self.model.cpu()
+            self.model.load_state_dict(torch.load(filepath))
+            self.model = self.model.cuda()
+        else:
+            self.model.load_state_dict(torch.load(filepath, map_location=lambda storage, loc: storage))
+        return self
+
+
+class Retina(ModelParallel):
     def __init__(self, architecture_config, training_config, callbacks_config, train_mode=False):
         """
         """
@@ -55,38 +129,6 @@ class Retina(Model):
         outputs = {'boxes_prediction': boxes,
                    'labels_prediction': labels}
         return outputs
-
-    def _fit_loop(self, data):
-        X = data[0]
-        targets_tensors = data[1:]
-
-        if torch.cuda.is_available():
-            X = Variable(X).cuda()
-            targets_var = []
-            for target_tensor in targets_tensors:
-                targets_var.append(Variable(target_tensor).cuda())
-        else:
-            X = Variable(X)
-            targets_var = []
-            for target_tensor in targets_tensors:
-                targets_var.append(Variable(target_tensor))
-
-        self.optimizer.zero_grad()
-        outputs_batch = self.model(X)
-        partial_batch_losses = {}
-
-        if len(self.output_names) == 1:
-            for (name, loss_function, weight), target in zip(self.loss_function, targets_var):
-                batch_loss = loss_function(outputs_batch, target) * weight
-        else:
-            for (name, loss_function, weight), output, target in zip(self.loss_function, outputs_batch, targets_var):
-                partial_batch_losses[name] = loss_function(output, target) * weight
-            batch_loss = sum(partial_batch_losses.values())
-        partial_batch_losses['sum'] = batch_loss
-        batch_loss.backward()
-        self.optimizer.step()
-
-        return partial_batch_losses
 
     def set_model(self):
         self.model = RetinaNet(encoder_depth=self.encoder_depth,
