--- src/steppy/preprocessing/misc.py
@@ -1,31 +1,0 @@
-from sklearn.externals import joblib
-
-from ..base import BaseTransformer
-
-
-class XYSplit(BaseTransformer):
-    def __init__(self, x_columns, y_columns):
-        self.x_columns = x_columns
-        self.y_columns = y_columns
-
-    def transform(self, meta, train_mode):
-        X = meta[self.x_columns].values
-        if train_mode:
-            y = meta[self.y_columns].values
-        else:
-            y = None
-
-        return {'X': X,
-                'y': y}
-
-    def load(self, filepath):
-        params = joblib.load(filepath)
-        self.columns_to_get = params['x_columns']
-        self.target_columns = params['y_columns']
-        return self
-
-    def save(self, filepath):
-        params = {'x_columns': self.x_columns,
-                  'y_columns': self.y_columns
-                  }
-        joblib.dump(params, filepath)
